reorganize files and add entrypoint
- I wanted to place test files under the test folder, but there was an error when import modules because my directory's name has '-'.
