+++ src/analysis/check.sh
