+++ src/analysis/reformat.sh
@@ -1,0 +1,2 @@
+poetry run black *.py
+poetry run isort --overwrite-in-place *.py
